requirements.txt
absl-py==0.7.1               
astor==0.7.1               
atari-py==0.2.3               
atomicwrites==1.3.0               
attrs==19.1.0              
backcall==0.1.0               
bleach==3.1.0               
certifi==2019.3.9            
cffi==1.12.3              
chardet==3.0.4               
Click==7.0                 
tfp-nightly==0.8.0.dev20190704             
cycler==0.10.0              
decorator==4.4.0               
defusedxml==0.6.0               
dopamine-rl==2.0.3               
EasyProcess==0.2.7               
entrypoints==0.3                 
enum34==1.1.6               
future==0.17.1              
gast==0.2.2               
gevent==1.4.0               
gin-config==0.1.3               
google-pasta==0.1.7               
greenlet==0.4.15              
grpcio==1.20.1              
gym==0.13.0
h5py==2.9.0               
idna==2.8                 
imageio==2.4.0               
importlib-metadata==0.17                
ipykernel==5.1.1               
ipython==7.5.0               
ipython-genutils==0.2.0               
ipywidgets==7.4.2               
jedi==0.13.3              
Jinja2==2.10.1              
joblib==0.13.2              
jsonschema==3.0.1               
jupyter==1.0.0               
jupyter-client==5.2.4               
jupyter-console==6.0.0               
jupyter-core==4.4.0               
Keras-Applications==1.0.8               
Keras-Preprocessing==1.0.9               
kiwisolver==1.1.0               
Markdown==3.1                 
MarkupSafe==1.1.1               
matplotlib==3.1.0               
mistune==0.8.4               
mock==3.0.4               
more-itertools==7.0.0               
nbconvert==5.5.0               
nbformat==4.4.0               
notebook==5.7.8               
numpy==1.16.3              
opencv-python==4.1.0.25            
opt-einsum==2.3.2               
packaging==19.0                
pandas==0.24.2              
pandocfilters==1.4.2               
parso==0.4.0               
pexpect==4.7.0               
pickleshare==0.7.5               
Pillow==6.0.0               
pip==19.1.1              
pkg-resources==0.0.0               
pluggy==0.12.0              
prometheus-client==0.7.0               
prompt-toolkit==2.0.9               
protobuf==3.7.1               
ptan==0.4                 
ptyprocess==0.6.0               
py==1.8.0               
pycparser==2.19                
pyglet==1.3.2               
Pygments==2.4.2               
pyparsing==2.4.0               
pyrsistent==0.15.2              
pytest==4.6.2               
python-dateutil==2.8.0               
pytz==2019.1              
PyVirtualDisplay==0.2.4               
pyzmq==18.0.1              
qtconsole==4.5.1               
requests==2.21.0              
scikit-learn==0.21.2              
scipy==1.2.1               
Send2Trash==1.5.0               
setuptools==41.0.1              
six==1.12.0              
sklearn==0.0                 
tb-nightly==1.14.0a20190614
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-agents-nightly==0.2.0.dev20190528
tf-estimator-nightly==1.14.0.dev2019070401
tf-nightly==1.15.0.dev20190704

torch==1.0.1.post2         
torchvision==0.2.2.post3         
tornado==6.0.2               
tqdm==4.32.1              
traitlets==4.3.2               
urllib3==1.24.3              
vizdoom==1.1.7               
wcwidth==0.1.7               
webencodings==0.5.1               
websocket==0.2.1               
websocket-client==0.56.0              
Werkzeug==0.15.2              
wheel==0.33.1              
widgetsnbextension==3.4.2               
wrapt==1.11.2              
zipp==0.5.1 
